chore(tags): code cleanup and refactoring
