+++ next-tavla/src/Admin/scenarios/BoardList/components/Row/index.tsx
@@ -1,13 +1,14 @@
 import { IconButton } from '@entur/button'
 import { CopyIcon, EditIcon } from '@entur/icons'
-import { TSettings } from 'types/settings'
 import { useToast } from '@entur/alert'
 import { useRouter } from 'next/router'
 import classes from './styles.module.css'
 import { useEffect, useState } from 'react'
 import { Cell } from 'Admin/scenarios/BoardList/components/Cell'
+import { Tags } from 'Admin/scenarios/Tags'
+import { Board } from 'types/board'
 
-function Row({ board }: { board: { id: string; settings?: TSettings } }) {
+function Row({ board }: { board: Board }) {
     const { addToast } = useToast()
     const router = useRouter()
     const [link, setLink] = useState('')
@@ -35,7 +36,10 @@ function Row({ board }: { board: { id: string; settings?: TSettings } }) {
                     <CopyIcon />
                 </IconButton>
             </Cell>
-            <Cell>
+            <Cell className={classes.tags}>
+                <Tags boardId={board.id} boardTags={board.settings?.tags} />
+            </Cell>
+            <Cell className={classes.options}>
                 <IconButton aria-label="Rediger tavle" onClick={editBoard}>
                     <EditIcon />
                 </IconButton>
@@ -44,4 +48,22 @@ function Row({ board }: { board: { id: string; settings?: TSettings } }) {
     )
 }
 
-export { Row }
+function TableHeader() {
+    return (
+        <div className={classes.tableHead}>
+            <div className={classes.tableRow}>
+                <div className={classes.headerCell}>Navn på tavle</div>
+                <div className={classes.headerCell}>Link</div>
+                <div className={classes.headerCell}>Tags</div>
+
+                <div
+                    className={`${classes.headerCell} ${classes.textCenteredCell}`}
+                >
+                    <div>Valg</div>
+                </div>
+            </div>
+        </div>
+    )
+}
+
+export { Row, TableHeader }
