--- next-tavla/src/Admin/scenarios/TableHeader/styles.module.css
@@ -1,21 +1,0 @@
-.tableHead {
-    padding: 0.5rem 0;
-    border-bottom: var(--colors-blues-blue20) solid;
-}
-
-.tableRow {
-    display: grid;
-    grid-template-columns: 3fr 5fr 3fr 5rem;
-}
-
-.tableHeaderCell {
-    display: flex;
-    align-items: center;
-    justify-content: flex-start;
-    font-weight: 500;
-    color: var(--colors-brand-lavender);
-}
-
-.tableHeaderOptionsCell {
-    justify-content: center;
-}
